backend/requirements.txt
# ── Framework web principal ──────────────────────────────────────
fastapi==0.115.6

# ── Servidor ASGI ────────────────────────────────────────────────
uvicorn==0.32.1

# ── ORM y base de datos ──────────────────────────────────────────
sqlalchemy==2.0.36
alembic==1.14.0

# ── Driver PostgreSQL (compatible con Python 3.13 en Apple Silicon)
psycopg[binary]==3.2.4

# ── Validación de datos ──────────────────────────────────────────
pydantic==2.10.3
pydantic-settings==2.6.1

# ── Variables de entorno ─────────────────────────────────────────
python-dotenv==1.0.1

# ── Autenticación JWT ────────────────────────────────────────────
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1

# ── Subida de archivos ───────────────────────────────────────────
python-multipart==0.0.12